# server/server/swagger.json
{
	"swagger": "2.0",
	"info": {
		"title": "ATLP-Blog API",
		"description": "API developed by using NodeJS and MongoDB",
		"version": "1.0.0",
		"contact": {
			"name": "NYAKAMWE Aimable",
			"email": "aimable.nyakamwe@yahoo.com",
			"url":"web.com"
		}
		
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"name": "DEV"
		}
	],
	"schemes":["http", "https"],
	"security": [{
		"bearerAuth": []
	  }],
	"paths": {
...
		"/api/messages/{id}": {
			"get": {
				"security": [{
					"bearerAuth": []
				  }],
				"tags": ["Messages | Contacts"],
				"summary": "Get Single message",
				"description": "get single message",
				"operationId": "get single message",
				
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "Add the token in the format 'Bearer token'",
						"required": true,
						"type": "string",
						"default": "Bearer "
					},
					{
						"name": "messageId",
						"in": "path",
						"type": "string",
						"description": "Message Id",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "success"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"delete": {
				"security": [{
					"bearerAuth": []
				  }],
				"tags": ["Messages | Contacts"],
				"summary": "delete message",
				"description": "delete message",
				"operationId": "delete message",
				
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "Add the token in the format 'Bearer token'",
						"required": true,
						"type": "string",
						"default": "Bearer "
					},
					{
						"name": "messageId",
						"in": "path",
						"type": "string",
						"description": "Message Id",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "successfull deleted"
					},
					"404": {
						"description": "contact not found"
					}
				}
			}
		},
...
		"/api/posts/{id}/comment": {
			"get": {
				"tags": ["Comments"],
				"summary": "Get all comments",
				"description": "get all comments",
				"operationId": "get all comments",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"type": "string",
						"description": "Post Id",
						"required": true
					}
				],
				"responses": {
					"201": {
						"description": "success"
					},
					"404": {
						"description": "not found"
					}
				}
				
			},
			"post": {
				"security": [{
					"bearerAuth": []
				  }],
				"tags": ["Comments"],
				"summary": "Post a new comment",
				"description": "New comment",
				"operationId": "New comment",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "Add the token in the format 'Bearer token'",
						"required": true,
						"type": "string",
						"default": "Bearer "
					},
					{
						"name": "id",
						"in": "path",
						"type": "string",
						"description": "Post Id",
						"required": true
					},
					{
						"in": "body",
						"name": "body",
						"description": "Resource payload",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"comment": {
									"type": "string"
								}
							},
							"required": ["comment"]
						}
					}
				],
				
				"responses": {
					"201": {
						"description": "successfully commented!"
					},
					"401": {
						"description": "you need to login"
					}
				}
			}
		},
...
	}
}